Add interactive mode

+++ cmd/scaffold.go
@@ -1,6 +1,8 @@
 package cmd
 
 import (
+	"fmt"
+	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -11,6 +13,7 @@ import (
 	"github.com/amido/stacks-cli/pkg/scaffold"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	"gopkg.in/yaml.v2"
 )
 
 var (
@@ -30,6 +33,8 @@ func init() {
 	// - options
 	var cmdlog bool
 	var dryrun bool
+	var saveConfig bool
+	var nocleanup bool
 
 	// - project settings
 	var project_name string
@@ -106,6 +111,8 @@ func init() {
 
 	scaffoldCmd.Flags().BoolVar(&cmdlog, "cmdlog", false, "Specify if commands should be logged")
 	scaffoldCmd.Flags().BoolVar(&dryrun, "dryrun", false, "Perform a dryrun of the CLI. No changes will be made on disk")
+	scaffoldCmd.Flags().BoolVar(&saveConfig, "save", false, "Save the the configuration from interactive or command line settings. Has no effect when using a configuration file.")
+	scaffoldCmd.Flags().BoolVar(&nocleanup, "nocleanup", false, "If set, do not perform cleanup at the end of the scaffolding")
 
 	// Bind the flags to the configuration
 
@@ -121,16 +128,14 @@ func init() {
 	viper.BindPFlag("project.sourcecontrol.url", scaffoldCmd.Flags().Lookup("sourcecontrolurl"))
 	viper.BindPFlag("project.sourcecontrol.ref", scaffoldCmd.Flags().Lookup("sourcecontrolref"))
 	viper.BindPFlag("project.settingsfile", scaffoldCmd.Flags().Lookup("projectsettingsfile"))
+	viper.BindPFlag("project.cloud.region", scaffoldCmd.Flags().Lookup("cloudregion"))
+	viper.BindPFlag("project.cloud.group", scaffoldCmd.Flags().Lookup("cloudgroup"))
 
 	viper.BindPFlag("settingsfile", scaffoldCmd.Flags().Lookup("settingsfile"))
 
 	viper.BindPFlag("pipeline", scaffoldCmd.Flags().Lookup("pipeline"))
 
-	viper.BindPFlag("platform.type", scaffoldCmd.Flags().Lookup("platformtype"))
-
 	viper.BindPFlag("cloud.platform", scaffoldCmd.Flags().Lookup("cloud"))
-	viper.BindPFlag("cloud.region", scaffoldCmd.Flags().Lookup("cloudregion"))
-	viper.BindPFlag("cloud.group", scaffoldCmd.Flags().Lookup("cloudgroup"))
 
 	viper.BindPFlag("business.company", scaffoldCmd.Flags().Lookup("company"))
 	viper.BindPFlag("business.domain", scaffoldCmd.Flags().Lookup("area"))
@@ -145,11 +150,23 @@ func init() {
 
 	viper.BindPFlag("options.cmdlog", scaffoldCmd.Flags().Lookup("cmdlog"))
 	viper.BindPFlag("options.dryrun", scaffoldCmd.Flags().Lookup("dryrun"))
+	viper.BindPFlag("options.save", scaffoldCmd.Flags().Lookup("save"))
+	viper.BindPFlag("options.nocleanup", scaffoldCmd.Flags().Lookup("nocleanup"))
 
 	viper.BindPFlag("interactive", scaffoldCmd.Flags().Lookup("interactive"))
 }
 
 func executeRun(ccmd *cobra.Command, args []string) {
+
+	// determine if the interactive option has been set
+	// if it has ask the user for input and then overwrite the configuration
+	// that has been specified on the command line with the values as given
+	// by the user
+	answers := config.Answers{}
+	err := answers.RunInteractive(&Config)
+	if err != nil {
+		App.Logger.Fatalf("Unable to perform interactive configuration: %s\n", err.Error())
+	}
 
 	// ensure that at least one project has been specified
 	if len(Config.Input.Project) == 1 && Config.Input.Project[0].Name == "" {
@@ -162,15 +179,65 @@ func executeRun(ccmd *cobra.Command, args []string) {
 		App.Logger.Fatalf("Specified pipeline is not supported: %s %v\n", Config.Input.Pipeline, pipeline.GetSupported())
 	}
 
+	// Attempt to save the configuration if the option has been set
+	// Determine if a configuration file has been used, if not save the settings from the
+	// InputObject to a yaml file so it can be used again without having to input all the details
+	// on the command line again
+	configFileUsed := viper.ConfigFileUsed()
+	if configFileUsed == "" && Config.Input.Options.SaveConfig {
+
+		// determine the path to the savedConfigFile
+		savedConfigFile := filepath.Join(Config.Input.Directory.WorkingDir, "stacks.yml")
+
+		App.Logger.Infof("Saving configuration file: %s\n", savedConfigFile)
+
+		// marshal the InputObject and write out the savedConfigFile
+		data, err := yaml.Marshal(&Config.Input)
+
+		if err != nil {
+			App.HandleError(err, "warn", "Problem converting configuration to YAML syntax")
+		}
+
+		err = ioutil.WriteFile(savedConfigFile, data, 0)
+
+		if err != nil {
+			App.HandleError(err, "warn", "Problem writing configuration to file")
+		}
+	}
+
 	// Call the scaffolding method
 	scaff := scaffold.New(&Config, App.Logger)
-	err := scaff.Run()
+	err = scaff.Run()
 	if err != nil {
 		App.Logger.Fatalf("Error running scaffold: %s", err.Error())
 	}
 
 	// call the func to set default values in the object
 	Config.SetDefaultValues()
+
+	// check that the framework binaries exist
+	missing := Config.Input.CheckFrameworks()
+	if len(missing) > 0 {
+
+		// iterate around the missing list
+		var list string
+		for _, item := range missing {
+
+			if item.Binary == "" {
+				list += fmt.Sprintf("Framework '%s' may have been misspelled because the command for this framework cannot be determined", item.Framework)
+			} else {
+				list += fmt.Sprintf("Command '%s' for the '%s' framework cannot be located. Is '%s' installed and in your PATH?", item.Binary, item.Framework, item.Binary)
+			}
+		}
+
+		// create the message to output to state that some binaries are missing
+		message := fmt.Sprintf(`Some of the commands required by the specified frameworks do not exist on your
+machine or the framework has been specified incorrectly.
+
+%s`, list)
+
+		App.Logger.Fatal(message)
+	}
 
 	// iterate around the projects that have been specified
 	for _, project := range Config.Input.Project {
@@ -183,10 +250,10 @@ func executeRun(ccmd *cobra.Command, args []string) {
 
 		// Clone the target repository into the temp directory
 		key := project.Framework.GetMapKey()
-		err := util.GitClone(
+		cloneDir, err := util.GitClone(
 			Config.Input.Stacks.GetSrcURL(key),
 			project.SourceControl.Ref,
-			project.Directory.TempDir,
+			Config.Input.Directory.TempDir,
 		)
 
 		// if an error occured getting the code then show an error and move onto the next project
@@ -196,12 +263,14 @@ func executeRun(ccmd *cobra.Command, args []string) {
 		}
 
 		// Read in the configuration file for the project
-		err = project.ReadSettings(project.Directory.TempDir, &Config)
+		err = project.ReadSettings(cloneDir, &Config)
 
 		if err != nil {
 			App.Logger.Errorf("Error reading settings from project settings: %s", err.Error())
 			continue
 		}
+
+		App.Logger.Infof("Read project settings from file: %s\n", project.SettingsFile)
 
 		// iterate around the init settings for the project
 		// all of these operations will occur in the temporary directory of the project
@@ -210,7 +279,7 @@ func executeRun(ccmd *cobra.Command, args []string) {
 			// log information based on the description in the settings file
 			App.Logger.Info(op.Description)
 
-			err = scaff.PerformOperation(op, &Config, &project, project.Directory.TempDir)
+			err = scaff.PerformOperation(op, &Config, &project, cloneDir)
 
 			if err != nil {
 				App.HandleError(err, "error", "Issue encountered performing init operation")
@@ -299,11 +368,15 @@ func executeRun(ccmd *cobra.Command, args []string) {
 	}
 
 	// Perform cleanup by removing the temporary directory
-	App.Logger.Info("Performing cleanup")
-	App.Logger.Infof(" - removing temporary directory: %s\n", Config.Input.Directory.TempDir)
-	err = os.RemoveAll(Config.Input.Directory.TempDir)
-	if err != nil {
-		App.Logger.Fatalf("Unable to remove temporary directory: %s", Config.Input.Directory.TempDir)
+	if Config.Input.Options.NoCleanup {
+		App.Logger.Warn("Cleanup has been disabled, please perform the cleanup manually")
+	} else {
+		App.Logger.Info("Performing cleanup")
+		App.Logger.Infof(" - removing temporary directory: %s\n", Config.Input.Directory.TempDir)
+		err = os.RemoveAll(Config.Input.Directory.TempDir)
+		if err != nil {
+			App.Logger.Fatalf("Unable to remove temporary directory: %s", Config.Input.Directory.TempDir)
+		}
 	}
 
 }
